requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Production
gunicorn==21.2.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[flask]==1.38.0
python-json-logger==2.0.7
smtplib-ssl==1.0.0

# Optional: AI/ML (uncomment if needed)
# openai==1.3.7

# Development (optional - add to dev-requirements.txt)
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
